+++ src/euring/__init__.py
@@ -13,7 +13,6 @@ Main features:
 
 from .__about__ import __version__
 from .converters import convert_euring2000_record, convert_euring_record
-from .decoders import EuringDecoder, euring_decode_record
 from .exceptions import EuringException, EuringParseException
 from .formats import (
     FORMAT_EURING2000,
@@ -52,10 +51,8 @@ from .utils import (
 __all__ = [
     "__version__",
     "EuringRecord",
-    "euring_decode_record",
     "convert_euring2000_record",
     "convert_euring_record",
-    "EuringDecoder",
     "TYPE_ALPHABETIC",
     "TYPE_ALPHANUMERIC",
     "TYPE_INTEGER",