+++ tests/smoke_test.py
@@ -14,13 +14,13 @@ def main():
     assert hasattr(euring, "__version__")
 
     # One minimal functional call
-    from euring import euring_decode_record
+    from euring import EuringRecord
 
-    record = euring_decode_record(
+    record = EuringRecord.decode(
         "DERA0CD...5206514ZZ1877018770N0ZUFF02U-----120719760----SV55+584200+01348000101030100202301739"
     )
     assert record.display_format == "EURING2000"
-    assert record["errors"] == {"record": [], "fields": []}
+    assert record.errors == {"record": [], "fields": []}
 
 
 if __name__ == "__main__":